+++ src/config.ts
@@ -15,7 +15,6 @@ export interface IServerConfig {
   hostExternalDomain: string;
   apiBaseUrl: string;
   namespace: string;
-  network: string;
   sectorsAllowed: Sector[];
   dbProvider: string;
   queueProvider: string;